% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AQuadtree-class.R
\name{area.QT}
\alias{area.QT}
\alias{area.QT,AQuadtree-method}
\title{area.QT AQuadtree-method}
\usage{
area.QT(obj, residual = TRUE, ...)

\S4method{area.QT}{AQuadtree}(obj, residual = FALSE, ...)
}
\arguments{
\item{obj}{an object of class AQuadtree.}

\item{residual}{logical; if TRUE cells marked as residual cells are included}

\item{...}{passed through.}
}
\value{
area of Quadtree grid cells in square meters
}
\description{
Method area.QT
}
\details{
Get the areas of the Quadtree grid cells in square meters
}
