% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BarcelonaCensusTracts}
\alias{BarcelonaCensusTracts}
\title{Census tract borders of Barcelona city in Catalonia.}
\format{
A SpatialPolygons object
}
\usage{
BarcelonaCensusTracts
}
\description{
A SpatiaPolygons object containing the Census tract borders of
Barcelona city in Catalonia.
}
\keyword{datasets}
