% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BarcelonaPop}
\alias{BarcelonaPop}
\title{Radomly created population points for Barcelona city in Catalonia.}
\format{A SpatialPointsDataFrame with 81359 rows and 2 attributes:
\describe{
  \item{age}{age the individual}
  \item{sex}{sex of the individual as a factor with levels: "man", "woman"}
}}
\usage{
BarcelonaPop
}
\description{
A dataset containing randomly created population for the Barcelona city
in Catalonia for the year 2018.
Population has been created randomly with the real distributions of census
tracts from the dataset dividing the total population by 20
(\url{https://www.bcn.cat/estadistica/catala/dades/tpob/pad/padro/a2018/edat/index.htm}).
}
\keyword{datasets}
