% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CharlestonCensusTracts}
\alias{CharlestonCensusTracts}
\title{Census tract borders of Charleston, SC MSA, USA.}
\format{A SpatialPolygons object with 117 polygons}
\usage{
CharlestonCensusTracts
}
\description{
A SpatiaPolygons object containing the Census tract borders of
Charleston, SC MSA, USA.
}
\keyword{datasets}
