% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CharlestonPop}
\alias{CharlestonPop}
\title{Radomly created population points for Charleston, SC MSA, USA.}
\format{
A SpatialPointsDataFrame with 54619 rows and 3 attributes:
\describe{
  \item{age}{group age the individual as a factor with
   levels: "under16", "16_65", "over65"}
  \item{sex}{sex of the individual as a factor with levels: "male", "female"}
  \item{origin}{origin of the individual as a factor with
   levels: "asian", "black", "hisp", "multi_ra", "white"}
}
}
\usage{
CharlestonPop
}
\description{
A dataset containing randomly created population for Charleston, SC MSA,
USA.
Population has been created randomly with distributions of census tracts
from the dataset Charleston1,	2000 Census Tract Data for Charleston, SC
MSA and counties
}
\seealso{
https://spatial.uchicago.edu/sample-data
}
\keyword{datasets}
