% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AQuadtree-class.R
\name{[,AQuadtree-method}
\alias{[,AQuadtree-method}
\alias{[}
\title{[ AQuadtree-method}
\usage{
\S4method{[}{AQuadtree}(x, i, j, ..., drop = TRUE)
}
\arguments{
\item{x}{an object of class AQuadtree.}

\item{i, j}{elements to extract.}

\item{...}{passed through.}

\item{drop}{passed on to [ indexing operator.}
}
\value{
An AQuadtree object with the selected subset of rows or columns from the input object.
}
\description{
Method [
}
\details{
Extract a part of a AQuadtree object
}
