% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AQuadtree-class.R
\name{merge,AQuadtree,data.frame-method}
\alias{merge,AQuadtree,data.frame-method}
\alias{merge}
\title{Merge an AQuadtree object with a data.frame}
\usage{
\S4method{merge}{AQuadtree,data.frame}(x, y)
}
\arguments{
\item{x}{an object of class AQuadtree.}

\item{y}{an object of class data.frame}
}
\value{
An AQuadtree object where the data is extended with the input data.frame
}
\description{
Method merge.
}
\details{
Merges the AQuadtree object data with the data.frame on the columns "cellCode" and cellNum"
}
