% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AQuadtree-class.R
\name{plot,AQuadtree,missing-method}
\alias{plot,AQuadtree,missing-method}
\alias{plot}
\title{plot AQuadtree-method}
\usage{
\S4method{plot}{AQuadtree,missing}(x, ..., residual = TRUE, add = FALSE, col)
}
\arguments{
\item{x}{an object of class AQuadtree.}

\item{...}{passed through.}

\item{residual}{logical; if TRUE cells marked as residual cells are included}

\item{add}{logical. TRUE to add plot to the current existing plot}

\item{col}{default plotting color}
}
\value{
none
}
\description{
Method plot
}
\details{
Plot an object of class AQuadtree.
}
