% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AQuadtree-class.R
\name{show,AQuadtree-method}
\alias{show,AQuadtree-method}
\alias{show}
\title{show AQuadtree-method}
\usage{
\S4method{show}{AQuadtree}(object)
}
\arguments{
\item{object}{an object of class AQuadtree.}
}
\value{
A data.frame showing the information attributes contained in the AQuadtree object.
}
\description{
Method show.
}
\details{
Display the AQuadtree object
}
