% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AQuadtree-class.R
\name{spplot,AQuadtree-method}
\alias{spplot,AQuadtree-method}
\alias{spplot}
\title{spplot AQuadtree-method}
\usage{
\S4method{spplot}{AQuadtree}(obj, zcol = NULL, by.density = TRUE, residual = TRUE, ...)
}
\arguments{
\item{obj}{an object of class AQuadtree.}

\item{zcol}{character; attribute name(s) or column number(s) in attribute table}

\item{by.density}{logical; if TRUE cell values specified in zcol are divided by cell areas}

\item{residual}{logical; if TRUE cells marked as residual cells are included}

\item{...}{passed through.}
}
\value{
Creates a lattice plot of class "trellis" created with the spplot method in the sp package
}
\description{
Method spplot
}
\details{
Plots a AQuadtree object as a spatial object with its data
}
