% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AQuadtree-class.R
\name{summary,AQuadtree-method}
\alias{summary,AQuadtree-method}
\alias{summary}
\title{summary AQuadtree-method}
\usage{
\S4method{summary}{AQuadtree}(object, ...)
}
\arguments{
\item{object}{an object of class AQuadtree.}

\item{...}{passed through.}
}
\value{
An object of class "table" with summarising information in the AQuadtree input object
}
\description{
Method summary.
}
\details{
summarize information of an object of class AQuadtree
}
