% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AQuadtree-class.R
\name{writeOGR.QT}
\alias{writeOGR.QT}
\alias{writeOGR.QT,AQuadtree-method}
\title{writeOGR.QT AQuadtree-method}
\usage{
writeOGR.QT(obj, residual = TRUE, ...)

\S4method{writeOGR.QT}{AQuadtree}(obj, residual = FALSE, ...)
}
\arguments{
\item{obj}{an object of class AQuadtree.}

\item{residual}{logical; if TRUE cells marked as residual cells are included}

\item{...}{passed through.}
}
\value{
if verbose=TRUE, a list of information about the attempted write operation
}
\description{
this is a generic for which \emph{rgdal} package also provides function.
}
\details{
Method writeOGR.QT

Exports a AQuadtree object as a SpatialPolygonsDataFrame using the
OGR abstraction library
}
